chore(home/chart): refactor actions
