+++ packages/blockchain-wallet-v4-frontend/src/scenes/Home/PriceChart/Actions/selectors.ts
@@ -10,7 +10,6 @@ export const getData = createDeepEqualSelector(
 
     return {
       coinName,
-      coinTicker,
       cryptoCurrency
     }
   }
